requirements.txt
evernote3>=1.25.14
html2text>=2020.1.16
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0
click>=8.1.0
tqdm>=4.65.0
platformdirs>=4.0


# GUI
PySide6>=6.9.0
# pyqtdarktheme 不支持 Python 3.13，改用内置 QSS
oauth2>=1.9.0            # evernote3 隐式依赖
setuptools>=60.0         # 为 Python 3.12+ 提供 distutils 兼容层
pytest-qt>=4.4.0
